+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changes in version 0.3.3
 * Added more Oxygen icons.
 * Added icon for *sharedcontent* plugin of *django-fluent-contents*.
 * Fixed ``KeyError`` when a model has add support, but no edit support.
+* Fixed icon layout when a model has no permissions to add/edit.
 
 
 Changes in version 0.3.2